chore: refactoring schemas and removing middleware
