+++ packages/database/src/schemas/index.ts
@@ -1,0 +1,8 @@
+export * from "./course";
+export * from "./organization";
+export * from "./school";
+export * from "./season";
+export * from "./session";
+export * from "./shipment";
+export * from "./user";
+export * from "./video";
